+++ vocatest/src/main/java/com/example/vocatest/controller/UserController.java
@@ -50,8 +50,6 @@ public class UserController {
             session.invalidate();
         }
 
-        
-
         return ResponseEntity.ok("Logged out successfully");
     }
