+++ vocatest/src/main/java/com/example/vocatest/service/UserService.java
@@ -16,8 +16,6 @@ import java.util.List;
 @Service
 public class UserService extends DefaultOAuth2UserService {
 
-
-  
     private final UserRepository userRepository;
 
     public UserService(UserRepository userRepository) {
